Reno: Faster recovery from a simple random loss
Simple lost detection: when the recovery episode ends
and the duration of current recovery episode is
short than a given threshold, currently defined by
2 x min_rtt. This means there was a small loss
recovered very quickly, indicating this has a
low probability of congestion but a simple packet lost.
In this case we restore previous cwnd/ssthresh
to resume faster after recovery.

This gives a good performance improvement in case of high
random loss, not by congestion. Test result shows that
it can improve more than 40% in case of 1% random loss.